successfully reading and interpretting as video frames, however I changed the coded to use 32 bis for xPos and yPos in encoded pixel. need to change that back

+++ hdr/readDepthFrame.hpp
@@ -7,8 +7,8 @@
 
 typedef struct{
     float diff;
-    int16_t xPos;
-    int16_t yPos;
+    uint32_t xPos;
+    uint32_t yPos;
 }DeltaEncodedPixel;
